better variant for double test macro

+++ checks/check_utils.h
@@ -6,7 +6,7 @@
 #include <check.h>
 #include <check_utils.h>
 
-#define ck_assert_dbl_eq(v1,v2,alow)  ck_assert_msg(fabs(v1-v2) < alow, "difference between %.10f and %.10f > %.10f", v1, v2, alow);
-#define ck_assert_lng_dbl_eq(v1,v2,alow)  ck_assert_msg(fabs(v1-v2) < alow, "difference between %.10f and %.10f > %.10f", v1, v2, alow);
+#define ck_assert_dbl_eq(v1,v2,alow)  ck_assert_msg(fabs((v1)-(v2)) < (alow), "difference between %.10f and %.10f > %.10f", (double)(v1), (double)(v2), (double)(alow));
+#define ck_assert_lng_dbl_eq(v1,v2,alow)  ck_assert_msg(fabs((v1)-(v2)) < (alow), "difference between %.10Lf and %.10Lf > %.10Lf", (long double)(v1), (long double)(v2), (long double)(alow));
 
 #endif //!__CHECK_UTILS__
